% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion.R
\name{RegionPlot}
\alias{RegionPlot}
\title{RegionPlot R6 Class}
\description{
RegionPlot R6 Class

RegionPlot R6 Class
}
\details{
The `RegionPlot` class provides functionality for creating and visualizing a genomic region plot using various tracks from the Gviz package. This class is particularly useful for visualizing specific regions of the human genome (hg38).
}
\examples{
\dontrun{
plotter <- RegionPlot$new("path_to_my_file.bam", "chr17", 7673300, 7675000)
plotter$plot(output_file = "regionPlot.pdf")
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bgFile}}{The path to the background file.}

\item{\code{chr}}{Chromosome name.}

\item{\code{from}}{Start position of the region.}

\item{\code{to}}{End position of the region.}

\item{\code{alignmentsTrack}}{AlignmentsTrack object.}

\item{\code{seqTrack}}{SequenceTrack object.}

\item{\code{ideoTrack}}{IdeogramTrack object.}

\item{\code{axisTrack}}{GenomeAxisTrack object.}

\item{\code{geneTrack}}{GeneRegionTrack object.}

\item{\code{txTrack}}{GeneRegionTrack object for knownGene.}

\item{\code{txTrackRef}}{GeneRegionTrack object for refGene.
Initialize the RegionPlot object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RegionPlot-new}{\code{RegionPlot$new()}}
\item \href{#method-RegionPlot-createTracks}{\code{RegionPlot$createTracks()}}
\item \href{#method-RegionPlot-plot}{\code{RegionPlot$plot()}}
\item \href{#method-RegionPlot-clone}{\code{RegionPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegionPlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-RegionPlot-new}{}}}
\subsection{Method \code{new()}}{
This method initializes the RegionPlot object with the
specified background file, chromosome, and genomic region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegionPlot$new(bgFile, chr, from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bgFile}}{The path to the background file.}

\item{\code{chr}}{Chromosome name.}

\item{\code{from}}{Start position of the region.}

\item{\code{to}}{End position of the region.
Create the necessary tracks for visualization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegionPlot-createTracks"></a>}}
\if{latex}{\out{\hypertarget{method-RegionPlot-createTracks}{}}}
\subsection{Method \code{createTracks()}}{
This method creates the AlignmentsTrack, SequenceTrack,
IdeogramTrack, GenomeAxisTrack, GeneRegionTrack, and
GeneRegionTrack objects for visualization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegionPlot$createTracks()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
Plot the tracks to a PDF file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegionPlot-plot"></a>}}
\if{latex}{\out{\hypertarget{method-RegionPlot-plot}{}}}
\subsection{Method \code{plot()}}{
This method plots the tracks to a PDF file for visualization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegionPlot$plot(output_file = "~/Downloads/regionPlot.pdf")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output_file}}{The path to the output PDF file. Default is "~/Downloads/regionPlot.pdf".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegionPlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RegionPlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegionPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
